+++ src/app/(app)/@modals/params.tsx
@@ -1,4 +1,5 @@
 export enum ModalParam {
   AddQuest = "add-quest",
+  QuestChains = "quest-chains",
   Stats = "stats",
 }